% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimandYLL-package.R
\docType{package}
\name{estimandYLL-package}
\alias{estimandYLL}
\alias{estimandYLL-package}
\title{estimandYLL: Years of Life Lost via the G-Formula}
\description{
Estimate differences in restricted expected residual lifetime under two
binary exposure scenarios, standardized to a stated baseline population.
See \code{\link[=estimate_yll]{estimate_yll()}} for the main estimator and interpretation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ShuntaroS/estimandYLL}
  \item \url{https://shuntaros.github.io/estimandYLL/}
  \item Report bugs at \url{https://github.com/ShuntaroS/estimandYLL/issues}
}

}
\author{
\strong{Maintainer}: Shuntaro Sato \email{shuntaro.sato.stat@gmail.com} (\href{https://orcid.org/0000-0002-6055-3649}{ORCID})

Authors:
\itemize{
  \item Shuntaro Sato \email{shuntaro.sato.stat@gmail.com} (\href{https://orcid.org/0000-0002-6055-3649}{ORCID})
}

}
\keyword{internal}
