% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimandYLL-package.R
\docType{package}
\name{estimandYLL-package}
\alias{estimandYLL}
\alias{estimandYLL-package}
\title{estimandYLL: Years of Life Lost via the G-Formula}
\description{
Estimates Years of Life Lost (YLL) and counterfactual life expectancy
under user-specified interventions on a binary exposure using the
parametric g-formula on the attained-age time scale. The main interface,
\code{\link[=estimand_yll]{estimand_yll()}}, makes the target population,
intervention, and result measure explicit.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ShuntaroS/estimandYLL}
  \item \url{https://shuntaros.github.io/estimandYLL/}
  \item Report bugs at \url{https://github.com/ShuntaroS/estimandYLL/issues}
}

}
\author{
\strong{Maintainer}: Shuntaro Sato \email{shuntaro.sato.stat@gmail.com} (\href{https://orcid.org/0000-0002-6055-3649}{ORCID})

}
\keyword{internal}
